Fix Warning
git-svn-id: svn://localhost/svn/uclinux-dist/trunk@2965 3e2a5524-39e8-4cb9-84a9-c49da69cba3a

+++ linux-2.6.x/drivers/video/bfin_adv7393fb.c
@@ -619,7 +619,7 @@ bfin_adv7393_fb_init(void)
   if(NULL == fbdev->fb_mem)
     {
       printk (KERN_ERR "FB: couldn't allocate dma buffer (%d bytes) \n",
-	      fbdev->fb_len);
+	     (u32) fbdev->fb_len);
       ret = -ENOMEM;
       goto out_5;
     }
